Move the test case to temporal-testing
